refactor: rename "main" folder to "components"
follow https://reactjs.org/docs/faq-structure.html#grouping-by-file-type
